+++ Code/base/xmake.lua
@@ -1,5 +1,5 @@
 
-target("Base")
+target("BaseLib")
     add_configfiles("BuildInfo.h.in")
     set_kind("static")
     set_group("common")