+++ Code/client/Services/Generic/OverlayClient.cpp
@@ -1,9 +1,15 @@
 #include <TiltedOnlinePCH.h>
 
 #include <OverlayRenderHandler.hpp>
-#include <Services/OverlayClient.h>
 
+#include <Services/OverlayClient.h>
 #include <Services/TransportService.h>
+
+#include <Events/CommandEvent.h>
+
+#include <Messages/SendChatMessageRequest.h>
+
+#include <World.h>
 
 OverlayClient::OverlayClient(TransportService& aTransport, TiltedPhoques::OverlayRenderHandler* apHandler)
     : TiltedPhoques::OverlayClient(apHandler), m_transport(aTransport)
@@ -34,18 +40,51 @@ bool OverlayClient::OnProcessMessageReceived(CefRefPtr<CefBrowser> browser, CefR
 #endif
 
         if (eventName == "connect")
-        {
-            std::string baseIp = eventArgs->GetString(0);
-            uint16_t port = eventArgs->GetInt(1) ? eventArgs->GetInt(1) : 10578;
-            m_transport.Connect(baseIp + ":" + std::to_string(port));
-        }
-        if (eventName == "disconnect")
-        {
-            m_transport.Close();
-        }
+            ProcessConnectMessage(eventArgs);
+        else if (eventName == "disconnect")
+            ProcessDisconnectMessage();
+        else if (eventName == "sendMessage")
+            ProcessChatMessage(eventArgs);
 
         return true;
     }
 
     return false;
+}
+
+void OverlayClient::ProcessConnectMessage(CefRefPtr<CefListValue> aEventArgs)
+{
+    std::string baseIp = aEventArgs->GetString(0);
+    if (baseIp == "localhost")
+    {
+        baseIp = "127.0.0.1";
+    }
+
+    uint16_t port = aEventArgs->GetInt(1) ? aEventArgs->GetInt(1) : 10578;
+    m_transport.Connect(baseIp + ":" + std::to_string(port));
+    // iAmAToken = aEventArgs->GeString(2);
+}
+
+void OverlayClient::ProcessDisconnectMessage()
+{
+    m_transport.Close();
+}
+
+void OverlayClient::ProcessChatMessage(CefRefPtr<CefListValue> aEventArgs)
+{
+    std::string contents = aEventArgs->GetString(0).ToString();
+    if (!contents.empty())
+    {
+        if (contents[0] == '/')
+        {
+            World::Get().GetRunner().Trigger(CommandEvent(std::move(String(contents))));
+        }
+        else
+        {
+            SendChatMessageRequest messageRequest;
+            messageRequest.ChatMessage = aEventArgs->GetString(0).ToString();
+            spdlog::debug("Received Message from UI and will send it to server: " + messageRequest.ChatMessage);
+            m_transport.Send(messageRequest);
+        }
+    }
 }