+++ Code/common/xmake.lua
@@ -1,5 +1,5 @@
 
-target("Common")
+target("CommonLib")
     add_configfiles("BuildInfo.h.in")
     set_kind("static")
     set_group("common")