+++ Code/components/crash_handler/CrashHandler.h
@@ -1,0 +1,4 @@
+#pragma once
+
+void InstallCrashHandler(bool aServer = false, bool aSkyrim = true);
+void UninstallCrashHandler();