+++ Code/encoding/Opcodes.h
@@ -41,6 +41,8 @@ enum ClientOpcode : unsigned char
     kRequestRespawn,
     kSyncExperienceRequest,
     kRequestEquipmentChanges,
+    kSendChatMessageRequest,
+    kTeleportCommandRequest,
     kClientOpcodeMax
 };
 
@@ -82,5 +84,7 @@ enum ServerOpcode : unsigned char
     kNotifyRespawn,
     kNotifySyncExperience,
     kNotifyEquipmentChanges,
+    kNotifyChatMessageBroadcast,
+    kTeleportCommandResponse,
     kServerOpcodeMax
 };