+++ Code/skyrim_ui/src/app/models/geo-location.ts
@@ -1,0 +1,16 @@
+export class GeoLocation {
+    countryCode: string;
+    continent: string;
+    country: string;
+
+    constructor(options: {
+        countryCode: string;
+        continent: string;
+        country;
+
+    }) {
+        this.countryCode = options.countryCode;
+        this.continent = options.continent;
+        this.country = options.country;
+    }
+}