+++ Code/skyrim_ui/src/app/services/server-list.service.ts
@@ -1,9 +1,10 @@
 import { Injectable } from '@angular/core';
 import { HttpClient } from '@angular/common/http';
 import { environment } from '../../environments/environment';
-import { Observable } from 'rxjs';
-import { ServerList } from '../models/server-list';
-import { map } from 'rxjs/operators';
+import { Observable, of } from 'rxjs';
+import { map, tap } from 'rxjs/operators';
+import { Server } from '../models/server';
+import { GeoLocation } from '../models/geo-location';
 
 @Injectable({
   providedIn: 'root'
@@ -12,14 +13,24 @@ export class ServerListService {
 
   constructor(private http: HttpClient) { }
 
-  public getServerList(): Observable<ServerList[]> {
-    return this.http.get<ServerList[]>(`${environment.urlProtocol}://list.${environment.baseUrl}/list`)
+  public getServerList(): Observable<Server[]> {
+    return this.http.get<Server[]>(`${environment.urlProtocol}://${environment.url}/list`)
       .pipe(
         map((data: any) => data.servers),
       );
   }
 
-  public getCountryForIp(ip: string) {
-    return this.http.get(`http://ip-api.com/json/${ip}?fields=country`);
+  public getInformationForIp(ip: string): Observable<GeoLocation> {
+    const cached = localStorage.getItem(`ip-cache:${ip}`);
+
+    if (cached) {
+      return of(JSON.parse(cached));
+    }
+
+    return this.http.get<GeoLocation>(`http://ip-api.com/json/${ip}?fields=continent,countryCode,country`).pipe(
+      tap((response) => localStorage.setItem(`ip-cache:${ip}`, JSON.stringify(response))),
+    );
   }
+
+
 }